fix(02.1): correct VadConfig Double→Float type mismatch and verify phase
WR-05 introduced vadProbabilityThreshold as Double but FluidAudio's
VadConfig expects Float. Fixed in ModelWarmupService and
TranscriptionService. Updated test assertion to reference shared
constant instead of stale 0.5 literal. Build succeeds, 59/59 tests pass.

+++ Dicticus/Dicticus/Services/TranscriptionService.swift
@@ -360,7 +360,7 @@ extension TranscriptionService {
             let models = try await AsrModels.downloadAndLoad(version: .v3)
             let manager = AsrManager(config: .default)
             try await manager.loadModels(models)
-            let vad = try await VadManager(config: VadConfig(defaultThreshold: Double(vadProbabilityThreshold)))
+            let vad = try await VadManager(config: VadConfig(defaultThreshold: Float(vadProbabilityThreshold)))
             return TranscriptionService(asrManager: manager, vadManager: vad)
         } catch {
             return nil
